テストを追加
git-svn-id: https://www.seasar.org/svn/sandbox/s2buri/trunk@556 1c266114-290a-0410-b1ce-a63319e86887

+++ buri-core-test/src/test/java/org/escafe/buri/dao/util/BuriDataUtilTest.java
@@ -93,6 +93,6 @@ public class BuriDataUtilTest extends S2TestCase {
 		assertTrue(plains.size() == 0);
 		path = new BuriPath("pkg.prces.act", "rpkg.rproces.ract");
 		realPath = pathUtil.getBuriPathFromRealPath(path);
-		assertEquals(fstID, realPath.getBuriPathId());
+		assertEquals(fstID, realPath.getBuriPathId().longValue());
 	}
 }
